refactor(hooks): create useResize and useSplit hooks
